// Teste_poui/src/app/financeiro/cad-taxa/models/taxa-cartao.model.ts
/**
 * @generated  poui-specialist v1.3
 * @author     Andre Costa <andre.andrelscosta@gmail.com>
 * @license    Uso permitido · redistribuição proibida sem autorização escrita
 * @see        https://github.com/Alscosta1973/poui-specialist
 */

export interface TaxaCartao {
  id: string;
  bandeira: string;
  tipo: string;
  parcelas: number;
  taxa: number;
  dtIni: string;
  dtFim: string;
}